requirements.txt
datasets
pyffx
transformers
torch 
torchvision 
torchaudio
accelerate>=0.26.0
tenseal 
opacus
spacy
stanza
flair
sentence_transformers
sdv
nltk
wandb
cryptography
pyyaml
faker
numpy
pandas